requirements.txt
GitPython>=3.1.27
pandas>=1.5.0
ipython~=8.0.1